requirements.txt
openai>=1.0.0
python-dotenv
streamlit
protobuf==3.20.3
requests